refactor: Raffle-Authentication 관심사 분리
